rename BooleanParameterInput to ParameterSwitch and move ParameterFormGrid into a separate file
